requirements.txt
alchemy-mock==0.4.3
alembic==1.4.3
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
attrs==20.2.0
awscli==1.18.147
backcall==0.2.0
black==20.8b1
botocore==1.18.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.3
dataclasses==0.7
decorator==4.4.2
Django==2.0.2
dnspython==1.16.0
docutils==0.15.2
eventlet==0.28.0
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.1
git-remote-codecommit==1.15.1
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
greenlet==0.4.17
heroku==0.1.4
httplib2==0.18.1
idna==2.10
ikp3db==1.4.1
importlib-metadata==1.7.0
iniconfig==1.0.1
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.4.2
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==8.5.0
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.4
parso==0.1.1
passlib==1.7.4
pathspec==0.8.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.6
protobuf==3.13.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.6.0
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.6
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.5
s3transfer==0.3.3
six==1.15.0
spoonacular==3.0
SQLAlchemy==1.3.19
stevedore==3.2.0
toml==0.10.1
traitlets==4.3.3
twarc==1.10.1
tweepy==3.9.0
twilio==6.45.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==16.2.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
